upi-fraud-detection-fullstack/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
lightgbm==4.5.0
pytest==8.3.3
